move special_case_ctx.h to coord dir
